Migrated cattle data with remote

+++ shared/src/main/java/com/pr656d/shared/domain/cattle/validator/CattleTagNumberValidatorUseCase.kt
@@ -53,6 +53,6 @@ class CattleTagNumberValidatorUseCase @Inject constructor(
             return
         }
 
-        isCattleExistWithTagNumberUseCase(tagNumber.toLong(), isCattleExistResult)
+        isCattleExistWithTagNumberUseCase(tagNumber, isCattleExistResult)
     }
 }
